% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{on_train_cross_validation_end}
\alias{on_train_cross_validation_end}
\title{On Training Cross Validation End}
\usage{
on_train_cross_validation_end(sess, index, average_error, num_samples,
  num_minibatches)
}
\arguments{
\item{sess}{Session instance on which to perform the operation}

\item{index}{(int): index of the current callback.}

\item{average_error}{(float): average error for the cross validation}

\item{num_samples}{(int): number of samples in cross validation}

\item{num_minibatches}{(int): number of minibatch in cross validation}
}
\description{
Callback that gets executed at the end of cross validation.
}
