% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{sess_minibatch_size_schedule}
\alias{sess_minibatch_size_schedule}
\title{Create a Minibatch Size Schedule}
\usage{
sess_minibatch_size_schedule(schedule, epoch_size = 1)
}
\arguments{
\item{schedule}{(int or list): if integer, this minibatch size will be used for the whole training.
In case of list of integers, the elements are used as the values for ``epoch_size`` samples.
If list contains pair, the second element is used as a value for (``epoch_size`` x first element) samples}

\item{epoch_size}{(int): number of samples as a scheduling unit.}
}
\description{
Creates a minibatch size schedule.
}
