% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{sess_training_session}
\alias{sess_training_session}
\title{Create Training Session Object}
\usage{
sess_training_session(trainer, mb_source, mb_size, model_inputs_to_streams,
  progress_frequency = NULL, max_samples = NULL, checkpoint_config = NULL,
  cv_config = NULL, test_config = NULL)
}
\arguments{
\item{trainer}{(Trainer): trainer}

\item{mb_source}{(MinibatchSource): minibatch source used for training}

\item{mb_size}{(minibatch_size_schedule): minibatch schedule for training}

\item{model_inputs_to_streams}{(dict): mapping between input variables and input streams}

\item{progress_frequency}{(int): frequency in samples for aggregated progress printing}

\item{max_samples}{(int): maximum number of samples used for training}

\item{checkpoint_config}{(~CheckpointConfig): checkpoint configuration}

\item{cv_config}{(~CrossValidationConfig): cross validation configuration}

\item{test_config}{(~TestConfig): test configuration}
}
\description{
A factory function to create a training session object.
}
